Add support for match guards
This abstracts out the variable binding so that it can be used in
multiple places. If there's a pre-existing guard on the match arm, we
wrap that guard in binding generation, and then have:
<literal guard> && <existing guard>

This also cleans up some code on the way, making more effective use of
the quote macros.