Generate literals via formatting instead of shifts
The previous implementation of mask generation used bit-shifts and
subtraction to generate the literals. This ended up implicitly using
32-bit literals for mask generation, which failed when dealing with
masks for 64-bit or 128-bit types.

This could have been fixed via making the masks use u128, but it's more
convenient to simply generate the appropriate mask in binary using
formatting directives. This will scale naturally if Rust ever adds a
u256 type or larger :3 and is nicer because it doesn't make an arbitrary
decision.

Fixes #6